Add simple search, only works at a fixed depth for now.
This commit adds the Negamax sekelton function and a search() function
that calls Negamax to find the score of a given position.

With this commit, it is already possible to connect Gargantua to a GUI
and play using the fixed depth setting. However, do not use depth
values greater than 7, or else the engine will take forever to find
the best move.